app-engine-pipeline/requirements.txt
Flask==1.0.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud-language==1.1.0
google-cloud-speech==0.36.0
google-cloud-texttospeech==0.2.0
google-cloud-vision==0.34.0
google-cloud-automl==0.1.1
googlemaps==3.0.2
gunicorn==19.9.0
pandas==0.23.4
